Removed cache ops & moved functionality into the actual caches. Cleaned up some of the code. Better operationalise some values.
