+++ src/main/scala/com/redbubble/util/fetch/FetchRunner.scala
@@ -2,19 +2,22 @@ package com.redbubble.util.fetch
 
 import com.redbubble.util.cache.{CacheKey, SimpleCache}
 import com.redbubble.util.fetch.TwitterFutureFetchMonadError.twitterFutureFetchMonadError
-import com.twitter.util.{Await, Future, FuturePool}
+import com.twitter.util.{Await, Duration, Future, FuturePool}
 import fetch._
 
-final case class FetchedObjectCache(underlying: SimpleCache) extends DataSourceCache {
-  override def get[A](k: DataSourceIdentity) = {
-    val valueFuture = underlying.get[A](CacheKey(k.toString)).rescue {
+final case class FetchedObjectCache(underlyingCache: SimpleCache) extends DataSourceCache {
+  private val WaitTimeout = Duration.fromSeconds(5)
+
+  override def get[A](k: DataSourceIdentity): Option[A] = {
+    // We rescue any failures from the underlying cache by returning None, which will cause a fetch from the datasource.
+    val cacheGet = underlyingCache.get[A](CacheKey(k.toString)).rescue {
       case _ => Future.value(None)
     }
-    Await.result(valueFuture)
+    Await.result(cacheGet, WaitTimeout)
   }
 
-  override def update[A](k: DataSourceIdentity, v: A) = {
-    underlying.put(CacheKey(k.toString()), v)
+  override def update[A](k: DataSourceIdentity, v: A): FetchedObjectCache = {
+    underlyingCache.put(CacheKey(k.toString()), v)
     this
   }
 }
@@ -23,8 +26,7 @@ final case class FetcherRunner(c: SimpleCache)(implicit fp: FuturePool) {
   private val cache = FetchedObjectCache(c)
 
   /**
-    * Runs this fetch as a `Future`, returning a `Future` containing the result.
+    * Runs this fetch using a `Future`, returning a `Future` containing the result.
     */
-  def runFetchF[T](fetch: Fetch[T]): Future[T] =
-    Fetch.run[Future](fetch, cache)(twitterFutureFetchMonadError)
+  def runFetchAsFuture[T](fetch: Fetch[T]): Future[T] = Fetch.run[Future](fetch, cache)(twitterFutureFetchMonadError)
 }
